+++ container/src/main/java/com/flipkart/poseidon/core/PoseidonServlet.java
@@ -19,6 +19,7 @@ package com.flipkart.poseidon.core;
 import com.flipkart.poseidon.api.Application;
 import com.flipkart.poseidon.api.Configuration;
 import com.flipkart.poseidon.exception.DataSourceException;
+import com.flipkart.poseidon.internal.ByteArrayDataType;
 import com.google.common.net.MediaType;
 import flipkart.lego.api.exceptions.BadRequestException;
 import flipkart.lego.api.exceptions.ElementNotFoundException;
@@ -186,13 +187,17 @@ public class PoseidonServlet extends HttpServlet {
         httpResponse.setStatus(statusCode);
         Object responseObj = poseidonResponse.getResponse();
         if (responseObj != null) {
-            String responseStr = "";
             if (responseObj instanceof String) {
-                responseStr = (String) responseObj;
+            	httpResponse.getWriter().println((String)responseObj);
+            } else if (responseObj instanceof ByteArrayDataType) {
+            	byte[] rawBytes = ((ByteArrayDataType)responseObj).getRawBytes();
+            	// we override default response meta-data as the data is raw bytes
+            	httpResponse.setContentType(MediaType.OCTET_STREAM.toString());
+            	httpResponse.setContentLength(rawBytes.length);
+            	httpResponse.getOutputStream().write(rawBytes);
             } else {
-                responseStr = configuration.getObjectMapper().writeValueAsString(responseObj);
+            	httpResponse.getWriter().println(configuration.getObjectMapper().writeValueAsString(responseObj));
             }
-            httpResponse.getWriter().println(responseStr);
         }
     }