+++ service-clients-core/src/main/java/com/flipkart/poseidon/serviceclients/ServiceContext.java
@@ -16,8 +16,9 @@
 
 package com.flipkart.poseidon.serviceclients;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentLinkedQueue;
 
 /**
  * This context is used to pass information specific to the service clients
@@ -40,32 +41,55 @@ public class ServiceContext {
         }
     };
 
+    private static final ThreadLocal<Boolean> isDebug = ThreadLocal.withInitial(() -> false);
+
+    private static final ThreadLocal<Map<String, List<ServiceDebug>>> debugResponses = ThreadLocal.withInitial(ConcurrentHashMap::new);
+    private static final ThreadLocal<Map<String, Queue<String>>> collectedHeaders = ThreadLocal.withInitial(HashMap::new);
+
     /**
      * initialize an empty service context, it will cleanup previous value of the threadlocal if used in a threadpool
+     * @param responseHeadersToCollect
      */
-    public static void initialize() {
+    public static void initialize(List<String> responseHeadersToCollect) {
         if (isImmutable.get()) {
             throw new UnsupportedOperationException();
         }
 
         context.remove();
         isImmutable.set(false);
+
+        if (responseHeadersToCollect == null || responseHeadersToCollect.isEmpty()) {
+            return;
+        }
+
+        responseHeadersToCollect.forEach(header -> {
+            collectedHeaders.get().put(header.toLowerCase(), new ConcurrentLinkedQueue<>());
+        });
     }
 
     /**
      * initialize a new service context with the given context, it will cleanup previous value of
      * the threadlocal if it is being used in a threadpool
      *
-     * @param ctxt Context map
+     * @param state State of the previous thread's context
      */
-    public static void initialize(Map<String, Object> ctxt) {
+    public static void initialize(ServiceContextState state) {
         if (isImmutable.get()) {
             throw new UnsupportedOperationException();
         }
 
         context.remove();
-        context.get().putAll(ctxt);
+        context.get().putAll(state.ctxt);
 
+        if (state.debug) {
+            isDebug.set(state.debug);
+
+            debugResponses.remove();
+            debugResponses.set(state.serviceResponses);
+        }
+
+        collectedHeaders.remove();
+        collectedHeaders.set(state.collectedHeaders);
         isImmutable.set(false);
     }
 
@@ -82,6 +106,44 @@ public class ServiceContext {
         }
 
         context.get().put(key, value);
+    }
+
+    /**
+     * Enables the collection of service response in ServiceContext
+     */
+    public static void enableDebug() {
+        if (isImmutable.get()) {
+            throw new UnsupportedOperationException();
+        }
+
+        isDebug.set(true);
+    }
+
+    /**
+     * Disables the collection of service response in ServiceContext
+     */
+    public static void disableDebug() {
+        if (isImmutable.get()) {
+            throw new UnsupportedOperationException();
+        }
+
+        isDebug.set(false);
+    }
+
+    public static boolean isDebug() {
+        return isDebug.get();
+    }
+
+    public static void addDebugResponse(String key, ServiceDebug response) {
+        if (!isDebug()) {
+            return;
+        }
+
+        debugResponses.get().computeIfAbsent(key, k -> new ArrayList<>()).add(response);
+    }
+
+    public static Map<String, List<ServiceDebug>> getDebugResponses() {
+        return debugResponses.get();
     }
 
     /**
@@ -109,5 +171,27 @@ public class ServiceContext {
     public static void shutDown() {
         context.remove();
         isImmutable.remove();
+        isDebug.remove();
+        debugResponses.remove();
+    }
+
+    static Map<String, Queue<String>> getCollectedHeaders() {
+        return collectedHeaders.get();
+    }
+
+    public static List<String> getCollectedHeaders(String header) {
+        return Optional.ofNullable(getCollectedHeaders())
+                .map(m -> m.get(header.toLowerCase()))
+                .map(ArrayList::new)
+                .orElseGet(null);
+    }
+
+    public static ServiceContextState getState() {
+        ServiceContextState state = new ServiceContextState();
+        state.ctxt = getContextMap();
+        state.collectedHeaders = getCollectedHeaders();
+        state.debug = isDebug();
+        state.serviceResponses = getDebugResponses();
+        return state;
     }
 }