+++ service-clients-core/src/main/java/com/flipkart/poseidon/serviceclients/ServiceResponseDecoder.java
@@ -23,14 +23,12 @@ import com.flipkart.poseidon.handlers.http.HttpResponseDecoder;
 import com.flipkart.poseidon.handlers.http.utils.StringUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.http.Header;
-import org.apache.http.HeaderIterator;
 import org.apache.http.HttpResponse;
 import org.slf4j.Logger;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
 /**
  * Created by mohan.pandian on 18/03/15.
@@ -43,24 +41,38 @@ public class ServiceResponseDecoder<T> implements HttpResponseDecoder<ServiceRes
     private final JavaType errorType;
     private final Logger logger;
     private final Map<String, Class<? extends ServiceClientException>> exceptions;
+    private final Map<String, Queue<String>> collectedHeaders;
+    private final Map<String, List<String>> localCollectedHeaders = new HashMap<>();
 
     public ServiceResponseDecoder(ObjectMapper objectMapper, JavaType javaType, JavaType errorType, Logger logger, Map<String, Class<? extends ServiceClientException>> exceptions) {
+        this(objectMapper, javaType, errorType, logger, exceptions, Collections.emptyMap());
+    }
+
+    public ServiceResponseDecoder(ObjectMapper objectMapper, JavaType javaType, JavaType errorType, Logger logger, Map<String, Class<? extends ServiceClientException>> exceptions, Map<String, Queue<String>> collectedHeaders) {
         this.objectMapper = objectMapper;
         this.javaType = javaType;
         this.errorType = errorType;
         this.logger = logger;
         this.exceptions = exceptions;
+        this.collectedHeaders = collectedHeaders;
     }
 
     private Map<String, String> getHeaders(HttpResponse httpResponse) {
         Map<String, String> headers = new HashMap<>();
-        HeaderIterator iterator = httpResponse.headerIterator();
-        if (iterator == null) {
+        Header[] responseHeaders = httpResponse.getAllHeaders();
+        if (responseHeaders == null || responseHeaders.length == 0) {
             return headers;
         }
-        while (iterator.hasNext()) {
-            Header header = iterator.nextHeader();
+        for (Header header : responseHeaders) {
             headers.put(header.getName(), header.getValue());
+            if (collectedHeaders.isEmpty()) {
+                continue;
+            }
+
+            String lowerCaseHeader = header.getName().toLowerCase();
+            if (collectedHeaders.containsKey(lowerCaseHeader)) {
+                localCollectedHeaders.computeIfAbsent(lowerCaseHeader, s -> new ArrayList<>()).add(header.getValue());
+            }
         }
         return headers;
     }
@@ -68,6 +80,7 @@ public class ServiceResponseDecoder<T> implements HttpResponseDecoder<ServiceRes
     @Override
     public ServiceResponse<T> decode(HttpResponse httpResponse) throws Exception {
         Map<String, String> headers = getHeaders(httpResponse);
+        collectedHeaders.forEach((k, v) -> Optional.ofNullable(localCollectedHeaders.get(k)).ifPresent(v::addAll));
         int statusCode = httpResponse.getStatusLine().getStatusCode();
         String statusCodeString = String.valueOf(statusCode);
         if (statusCode >= 200 && statusCode <= 299) {
@@ -144,7 +157,7 @@ public class ServiceResponseDecoder<T> implements HttpResponseDecoder<ServiceRes
     }
 
     @Override
-    public ServiceResponse<T> decode(InputStream is) {
-        throw new UnsupportedOperationException();
+    public ServiceResponse<T> decode(InputStream is) throws Exception {
+        return new ServiceResponse<T>(objectMapper.<T>readValue(is, javaType), null);
     }
 }