+++ CHANGELOG.md
@@ -1,9 +1,22 @@
-## 5.1.0 (WIP) (Mar 24, 2017)
+## 5.4.0-SNAPSHOT (Jul 4, 2017)
+- Throw original exceptions for service failures (de-serialization exceptions etc) without wrapping in other exceptions
+- Exposed an interface in Validator framework for user defined logic
+
+## 5.3.0 (June 27, 2017)
+- Ability to define and consume additional fields for an endpoint to be used as meta information
+
+## 5.2.0 (June 7, 2017)
+- Debug flag in APIs to list all service responses
+- Collect response headers from all upstream services for a particular call to the system.
+
+## 5.1.0 (May 25, 2017)
 - Upgrade Jetty version to v9.4.1.v20170120
 - Support incoming PATCH method calls
 - Support PATCH calls for service clients
 - Bumping Phantom to 3.4.0
 - HttpConnectionPool supports form encoded POST body
+- API reponse status metric names have http methods as well
+- Support for returning raw bytes in response
 
 ## 5.0.0 (Jan 2, 2017)
 - Included snapshot version fixes